Use ps in era5 reader and get water condensates

+++ tools/get_era5.py
@@ -30,7 +30,11 @@ if not os.path.isfile(plev_file):
 				'temperature',
 				'u_component_of_wind',
 				'v_component_of_wind',
-				'specific_humidity'
+				'specific_humidity',
+				'specific_cloud_liquid_water_content',
+				'specific_cloud_ice_water_content',
+				'specific_rain_water_content',
+				'specific_snow_water_content'
 			],
 			'pressure_level': [
 				'1'  , '2'  , '3'  ,
@@ -63,7 +67,8 @@ if not os.path.isfile(sfc_file):
 			'grid': grids[args.res],
 			'format': 'netcdf',
 			'variable': [
-				'mean_sea_level_pressure',
+				'surface_pressure',
+				'geopotential'
 			],
 			'year': args.time.format('YYYY'),
 			'month': args.time.format('MM'),
@@ -72,8 +77,12 @@ if not os.path.isfile(sfc_file):
 		},
 		sfc_file
 	)
+	os.system(f'cdo chname,z,zs {sfc_file} {sfc_file}.tmp')
+	os.system(f'mv {sfc_file}.tmp {sfc_file}')
 
 out_file = f'era5_{args.time.format("YYYYMMDDHH")}.nc'
 if not os.path.isfile(out_file):
-	os.system(f'cdo merge {plev_file} {sfc_file} {out_file}')
-	os.system(f'rm -f {plev_file} {sfc_file}')
+	os.system(f'cdo merge {plev_file} {sfc_file} {out_file}.tmp')
+	os.system(f'nccopy -d 6 {out_file}.tmp {out_file}')
+	if os.path.isfile(out_file):
+		os.system(f'rm -f {plev_file} {sfc_file}')
